Tighten crops-api after review: shared fixtures, integer chainId param, network-derived caveat
The OpenAPI agreement test imported its fixtures from the generator test
file, which made mocha load one suite as a side effect of another; the
fixtures now live in src/test/fixtures.ts. The chainId path parameter was
documented as a string while the response declares a number. The Sepolia
caveat in the spec description was literal prose that would have outlived
a move to mainnet; it is now derived from the attestation network
constant. Also: the Stamp type precedes its first use, the CLI no longer
takes an output directory nothing passes, and ProjectService is named as
such.
