+++ packages/crops-api/src/generateCropsSite.ts
@@ -36,6 +36,11 @@ export interface GeneratedFile {
   body: unknown
 }
 
+/** What every JSON file carries besides its own data. */
+type Stamp = Pick<GeneratorInput, 'commit' | 'generatedAt'> & {
+  attestations: CropsAttestationsMeta
+}
+
 /**
  * The whole API as a pure function of config and ledger, so it can be tested
  * with fixtures and regenerated byte-for-byte from the same commit.
@@ -56,16 +61,11 @@ export function generateCropsSite(input: GeneratorInput): GeneratedFile[] {
   return files.sort((a, b) => a.path.localeCompare(b.path))
 }
 
-type Stamp = Pick<GeneratorInput, 'commit' | 'generatedAt'> & {
-  attestations: CropsAttestationsMeta
-}
-
 function projectFiles(
   projects: CropsApiProject[],
   stamp: Stamp,
 ): GeneratedFile[] {
   return projects.flatMap((project) =>
-    // When id equals slug it is one file.
     [...new Set([project.id, project.slug])].map((key) => ({
       path: `v1/project/${key}.json`,
       body: { ...stamp, ...project },
