+++ packages/crops-api/src/openapi.ts
@@ -1,3 +1,4 @@
+import { CROPS } from '@l2beat/config'
 import { toJsonSchema, type Validator, v } from '@l2beat/validate'
 import {
   AddressesResponseSchema,
@@ -17,7 +18,7 @@ export interface PublishedRoute {
   path: string
   summary: string
   description: string
-  params: { name: string; description: string }[]
+  params: { name: string; type: 'string' | 'integer'; description: string }[]
   result: Validator<unknown>
   /** Set on lookups, where a missing file means "not reviewed". */
   notFound?: string
@@ -37,7 +38,13 @@ export const PUBLISHED_ROUTES: PublishedRoute[] = [
     summary: 'Everything about one protocol',
     description:
       'The crop evaluations of one protocol, with the reasoning behind each rating.',
-    params: [{ name: 'id', description: 'The project id or its slug.' }],
+    params: [
+      {
+        name: 'id',
+        type: 'string',
+        description: 'The project id or its slug.',
+      },
+    ],
     result: ProjectResponseSchema,
     notFound: 'L2BEAT has not reviewed this project.',
   },
@@ -49,9 +56,14 @@ export const PUBLISHED_ROUTES: PublishedRoute[] = [
     params: [
       {
         name: 'chainId',
+        type: 'integer',
         description: 'EIP-155 chain id, e.g. 1 for Ethereum.',
       },
-      { name: 'address', description: 'Lowercase 0x-prefixed address.' },
+      {
+        name: 'address',
+        type: 'string',
+        description: 'Lowercase 0x-prefixed address.',
+      },
     ],
     result: AddressResponseSchema,
     notFound: 'The address is not part of any reviewed protocol.',
@@ -80,6 +92,18 @@ function toPathRegex(template: string): RegExp {
   return new RegExp(`^${escaped}$`)
 }
 
+const ATTESTATION_NETWORK =
+  CROPS.eas.ATTESTATION_NETWORKS[CROPS.eas.ATTESTATION_NETWORK]
+
+/** Derived from the network constant so the caveat cannot outlive the testnet. */
+const ATTESTATION_NETWORK_SECTION = ATTESTATION_NETWORK.isTestnet
+  ? `## Attestations are on ${ATTESTATION_NETWORK.name}
+
+The attestation currently lives on the ${ATTESTATION_NETWORK.name} testnet (chain id ${ATTESTATION_NETWORK.chainId}); \`attestations.isTestnet\` in every response says so. It proves the set L2BEAT named, not that the ratings are attested: ratings change as protocols change and are served here without a transaction.`
+  : `## Attestations are on ${ATTESTATION_NETWORK.name}
+
+The attestation lives on ${ATTESTATION_NETWORK.name} (chain id ${ATTESTATION_NETWORK.chainId}). It proves the set L2BEAT named, not that the ratings are attested: ratings change as protocols change and are served here without a transaction.`
+
 export const OPENAPI_DESCRIPTION = `
 CROPS is L2BEAT's review of a protocol along four crops: censorship resistance, open source, privacy and security. Each crop gets a sentiment (good, warning, bad or neutral) and a status saying how far the review went. A protocol is in the garden when no crop is bad. Separately, the set of reviewed protocols is attested onchain with the Ethereum Attestation Service.
 
@@ -93,9 +117,7 @@ Every response is a static file generated from the L2BEAT repository, so it is s
 
 Address files are keyed by EIP-155 chain id and lowercase address, both in the path and in the \`addresses\` map of \`/v1/addresses.json\` (as \`chainId:address\`). Lowercase the address before you build the URL. A contract, a proxy implementation behind it, and a permission holder such as a governance multisig all resolve to the protocol. A shared contract lists every protocol that claims it, each once, with the name that protocol gives it.
 
-## Attestations are on Sepolia
-
-The attestation currently lives on the Sepolia testnet; \`attestations.isTestnet\` in every response says so. It proves the set L2BEAT named, not that the ratings are attested: ratings change as protocols change and are served here without a transaction.
+${ATTESTATION_NETWORK_SECTION}
 
 ## Verifying the set onchain
 
@@ -147,7 +169,7 @@ function toOperation(route: PublishedRoute) {
       in: 'path',
       required: true,
       description: param.description,
-      schema: { type: 'string' },
+      schema: { type: param.type },
     })),
     responses,
   }
